Move run_transaction into impl of FirestoreDb
