+++ src/db/mod.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::too_many_arguments)]
 
 mod get;
+use backoff::{future::retry, ExponentialBackoff};
 pub use get::*;
 
 mod create;
@@ -63,9 +64,10 @@ mod batch_simple_writer;
 pub use batch_simple_writer::*;
 
 use crate::errors::{
-    FirestoreError, FirestoreInvalidParametersError, FirestoreInvalidParametersPublicDetails,
+    FirestoreDatabaseError, FirestoreError, FirestoreInvalidParametersError,
+    FirestoreInvalidParametersPublicDetails,
 };
-use std::fmt::Formatter;
+use std::{fmt::Formatter, pin::Pin};
 
 #[derive(Clone)]
 pub struct FirestoreDb {
@@ -159,6 +161,89 @@ impl FirestoreDb {
             .await
             .ok();
         Ok(())
+    }
+
+    pub async fn run_transaction<T, F>(&self, func: F) -> Result<T, FirestoreError>
+    where
+        F: for<'a> Fn(
+            FirestoreDb,
+            &'a mut FirestoreTransaction,
+        )
+            -> Pin<Box<dyn futures::Future<Output = Result<T, FirestoreError>> + 'a>>,
+    {
+        // Perform our initial attempt. If this fails and the backend tells us we can retry,
+        // we'll try again with exponential backoff using the first attempt's transaction ID.
+        let transaction_id = {
+            let mut transaction = self.begin_transaction().await?;
+
+            let cdb = self.clone_with_consistency_selector(
+                FirestoreConsistencySelector::Transaction(transaction.transaction_id.clone()),
+            );
+
+            let ret_val = func(cdb, &mut transaction).await?;
+
+            let transaction_id = transaction.transaction_id.clone();
+
+            match transaction.commit().await {
+                Ok(_) => return Ok(ret_val),
+                Err(e) => match e {
+                    FirestoreError::DatabaseError(FirestoreDatabaseError {
+                        retry_possible: true,
+                        ..
+                    }) => {
+                        // Ignore; we'll try again
+                    }
+                    FirestoreError::DatabaseError(FirestoreDatabaseError {
+                        retry_possible: false,
+                        ..
+                    }) => {
+                        return Err(e);
+                    }
+                    e => return Err(e),
+                },
+            }
+
+            transaction_id
+        };
+
+        // We failed the first time. Now we must change the transaction mode to signal that we're retrying with the original transaction ID.
+        let result = retry(ExponentialBackoff::default(), || async {
+            let options = FirestoreTransactionOptions {
+                mode: FirestoreTransactionMode::ReadWriteRetry(transaction_id.clone()),
+            };
+            let mut transaction = self.begin_transaction_with_options(options).await?;
+
+            let cdb = self.clone_with_consistency_selector(
+                FirestoreConsistencySelector::Transaction(transaction.transaction_id.clone()),
+            );
+
+            let ret_val = func(cdb, &mut transaction).await?;
+
+            match transaction.commit().await {
+                Ok(_) => return Ok(ret_val),
+                Err(e) => match e {
+                    FirestoreError::DatabaseError(FirestoreDatabaseError {
+                        retry_possible: true,
+                        ..
+                    }) => {
+                        eprintln!("Got back retryable error: {e}");
+                        return Err(backoff::Error::transient(e));
+                    }
+                    FirestoreError::DatabaseError(FirestoreDatabaseError {
+                        retry_possible: false,
+                        ..
+                    }) => {
+                        return Err(backoff::Error::permanent(e));
+                    }
+                    e => return Err(backoff::Error::permanent(e)),
+                },
+            }
+        })
+        .await;
+
+        // TODO If the final result was Err, do we still need to call `rollback`?
+
+        result
     }
 
     #[inline]
